refactor: move shared logic into abstract class
